Finalizando o projeto!

+++ SAP_1/Controllers/CursoController.cs
@@ -28,47 +28,100 @@ namespace SAP_1.Controllers
             _service = service;
         }
 
-        [HttpGet]
-        public IActionResult Index()
+    public IActionResult Index()
+    {
+        return View();
+    }
+
+    [HttpPost]
+    public IActionResult GetCursos()
+    {
+        try
         {
-            List<Curso> cursos = _service.FindAll().ToList();
-            return View(cursos);
+            var draw = Request.Form["draw"].FirstOrDefault();
+            var start = Request.Form["start"].FirstOrDefault();
+            var length = Request.Form["length"].FirstOrDefault();
+            var sort = Request.Form["sort"].FirstOrDefault();
+            var searchValue = Request.Form["search[value]"].FirstOrDefault();
+            int pageSize = length != null ? Convert.ToInt32(length) : 0;
+            int skip = start != null ? Convert.ToInt32(start) : 0;
+            int recordsTotal = 0;
+            var cursos = _service.FindAll();
+            if (!string.IsNullOrEmpty(searchValue))
+            {
+                cursos = cursos.Where(c => c.DsCurso.Contains(searchValue)).ToList();
+            }
+            recordsTotal = cursos.Count();
+            var data = cursos.Skip(skip).Take(pageSize).Select(c => new {
+                idCurso = c.IdCurso,
+                dsCurso = c.DsCurso,
+                categoria = c.Categoria,
+                duracao = c.Duracao,
+                ativo = c.FgAtivo
+            }).ToList();
+
+            var jsonData = new 
+            { 
+                draw = draw,
+                recordsFiltered = recordsTotal,
+                recordsTotal = recordsTotal,
+                data = data 
+            };
+            return Json(jsonData);
         }
+        catch (System.Exception)
+        {
+            throw;
+        }
+    }
 
         [HttpGet]
         public IActionResult Criar()
         {
             ViewBag.StatusCurso = _status;
             ViewBag.CategoriaCurso = _categorias;
+
             return View();
         }
 
         [HttpPost]
         public IActionResult Criar(Curso curso)
         {
-            _service.Create(curso);
-            return RedirectToAction("Index");
+            if (curso != null)
+            {
+                _service.Create(curso);
+                return RedirectToAction("Index");
+            }
+
+            return NotFound();
         }
 
         [HttpGet]
         public IActionResult Remover(string idCurso)
         {
             Curso curso = _service.Find(new Curso { IdCurso = idCurso });
-            return View(curso);
+
+            return curso == null ? NotFound() : View(curso);
         }
 
         [HttpPost]
         public IActionResult Remover(Curso curso)
         {
-            _service.Delete(curso);
-            return RedirectToAction("Index");
+            if (curso != null)
+            {
+                _service.Delete(curso);
+                return RedirectToAction("Index");
+            }
+            return NotFound();
+
         }
 
         [HttpGet]
         public IActionResult Detalhar(string idCurso)
         {
             Curso curso = _service.Find(new Curso { IdCurso = idCurso });
-            return View(curso);
+
+            return curso == null ? NotFound() : View(curso);
         }
 
         [HttpGet]
@@ -77,7 +130,8 @@ namespace SAP_1.Controllers
             ViewBag.CategoriaCurso = _categorias;
             ViewBag.StatusCurso = _status;
             Curso curso = _service.Find(new Curso { IdCurso = idCurso });
-            return View(curso);
+
+            return curso == null ? NotFound() : View(curso);
         }
 
         [HttpPost]
